Added `tests` for `PART-4`
