+++ src/Neo.IO/Neo.IO.csproj
@@ -13,6 +13,7 @@
     <InternalsVisibleTo Include="Neo" />
     <InternalsVisibleTo Include="Neo.UnitTests" />
     <InternalsVisibleTo Include="Neo.Extensions" />
+    <InternalsVisibleTo Include="Neo.Extensions.Tests" />
   </ItemGroup>
 
   <ItemGroup>
