+++ tests/Neo.Extensions.Tests/Collections/UT_HashSetExtensions.cs
@@ -1,0 +1,75 @@
+// Copyright (C) 2015-2024 The Neo Project.
+//
+// UT_HashSetExtensions.cs file belongs to the neo project and is free
+// software distributed under the MIT software license, see the
+// accompanying file LICENSE in the main directory of the
+// repository or http://www.opensource.org/licenses/mit-license.php
+// for more details.
+//
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
+using Neo.IO.Caching;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Neo.Extensions.Tests.Collections
+{
+    [TestClass]
+    public class UT_HashSetExtensions
+    {
+        [TestMethod]
+        public void TestRemoveHashsetDictionary()
+        {
+            var a = new HashSet<int>
+            {
+                1,
+                2,
+                3
+            };
+
+            var b = new Dictionary<int, object>
+            {
+                [2] = null
+            };
+
+            a.Remove(b);
+
+            CollectionAssert.AreEqual(new int[] { 1, 3 }, a.ToArray());
+
+            b[4] = null;
+            b[5] = null;
+            b[1] = null;
+            a.Remove(b);
+
+            CollectionAssert.AreEqual(new int[] { 3 }, a.ToArray());
+        }
+
+        [TestMethod]
+        public void TestRemoveHashsetSet()
+        {
+            var a = new HashSet<int>
+            {
+                1,
+                2,
+                3
+            };
+
+            var b = new SortedSet<int>()
+            {
+                2
+            };
+
+            a.Remove(b);
+
+            CollectionAssert.AreEqual(new int[] { 1, 3 }, a.ToArray());
+
+            b.Add(4);
+            b.Add(5);
+            b.Add(1);
+            a.Remove(b);
+
+            CollectionAssert.AreEqual(new int[] { 3 }, a.ToArray());
+        }
+    }
+}
