+++ tests/Neo.UnitTests/UT_Helper.cs
@@ -52,75 +52,6 @@ namespace Neo.UnitTests
         }
 
         [TestMethod]
-        public void TestHexToBytes()
-        {
-            string nullStr = null;
-            _ = nullStr.HexToBytes().ToHexString().Should().Be(Array.Empty<byte>().ToHexString());
-            string emptyStr = "";
-            emptyStr.HexToBytes().ToHexString().Should().Be(Array.Empty<byte>().ToHexString());
-            string str1 = "hab";
-            Action action = () => str1.HexToBytes();
-            action.Should().Throw<FormatException>();
-            string str2 = "0102";
-            byte[] bytes = str2.HexToBytes();
-            bytes.ToHexString().Should().Be(new byte[] { 0x01, 0x02 }.ToHexString());
-        }
-
-        [TestMethod]
-        public void TestRemoveHashsetDictionary()
-        {
-            var a = new HashSet<int>
-            {
-                1,
-                2,
-                3
-            };
-
-            var b = new Dictionary<int, object>
-            {
-                [2] = null
-            };
-
-            a.Remove(b);
-
-            CollectionAssert.AreEqual(new int[] { 1, 3 }, a.ToArray());
-
-            b[4] = null;
-            b[5] = null;
-            b[1] = null;
-            a.Remove(b);
-
-            CollectionAssert.AreEqual(new int[] { 3 }, a.ToArray());
-        }
-
-        [TestMethod]
-        public void TestRemoveHashsetSet()
-        {
-            var a = new HashSet<int>
-            {
-                1,
-                2,
-                3
-            };
-
-            var b = new SortedSet<int>()
-            {
-                2
-            };
-
-            a.Remove(b);
-
-            CollectionAssert.AreEqual(new int[] { 1, 3 }, a.ToArray());
-
-            b.Add(4);
-            b.Add(5);
-            b.Add(1);
-            a.Remove(b);
-
-            CollectionAssert.AreEqual(new int[] { 3 }, a.ToArray());
-        }
-
-        [TestMethod]
         public void TestRemoveHashsetHashSetCache()
         {
             var a = new HashSet<int>
@@ -145,52 +76,6 @@ namespace Neo.UnitTests
             a.Remove(b);
 
             CollectionAssert.AreEqual(new int[] { 3 }, a.ToArray());
-        }
-
-        [TestMethod]
-        public void TestGetVersion()
-        {
-            // assembly without version
-
-            var asm = AppDomain.CurrentDomain.GetAssemblies()
-                .Where(u => u.FullName == "Anonymously Hosted DynamicMethods Assembly, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null")
-                .FirstOrDefault();
-            string version = asm?.GetVersion() ?? "";
-            version.Should().Be("0.0.0");
-        }
-
-        [TestMethod]
-        public void TestNextBigIntegerForRandom()
-        {
-            Random ran = new();
-            Action action1 = () => ran.NextBigInteger(-1);
-            action1.Should().Throw<ArgumentException>();
-
-            ran.NextBigInteger(0).Should().Be(0);
-            ran.NextBigInteger(8).Should().NotBeNull();
-            ran.NextBigInteger(9).Should().NotBeNull();
-        }
-
-        [TestMethod]
-        public void TestUnmapForIPAddress()
-        {
-            var addr = new IPAddress(new byte[] { 127, 0, 0, 1 });
-            addr.UnMap().Should().Be(addr);
-
-            var addr2 = addr.MapToIPv6();
-            addr2.UnMap().Should().Be(addr);
-        }
-
-        [TestMethod]
-        public void TestUnmapForIPEndPoin()
-        {
-            var addr = new IPAddress(new byte[] { 127, 0, 0, 1 });
-            var endPoint = new IPEndPoint(addr, 8888);
-            endPoint.UnMap().Should().Be(endPoint);
-
-            var addr2 = addr.MapToIPv6();
-            var endPoint2 = new IPEndPoint(addr2, 8888);
-            endPoint2.UnMap().Should().Be(endPoint);
         }
     }
 }
